TweetPolicyをControllerに適用

+++ app/Http/Controllers/TweetController.php
@@ -73,6 +73,8 @@ class TweetController extends Controller
      */
     public function edit(Tweet $tweet)
     {
+        $this->authorize('update', $tweet);
+
         $tags = Tag::all();
         $selectedTags = $tweet->tags->pluck('id')->all();
         return view('edit', [
@@ -91,6 +93,8 @@ class TweetController extends Controller
      */
     public function update(Request $request, Tweet $tweet)
     {
+        $this->authorize('update', $tweet);
+
         $tweet->update([
             'message' => $request->message
         ]);
@@ -107,6 +111,8 @@ class TweetController extends Controller
      */
     public function destroy(Tweet $tweet)
     {
+        $this->authorize('delete', $tweet);
+
         $tweet->tags()->detach();
         $tweet->delete();
         return redirect()->route('tweets.index');
